requirements-dev.txt
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==6.1.0
coverage==7.8.0
django-debug-toolbar==6.2.0
django-extensions==4.1
pre-commit==4.5.0
pylint-django==2.6.1
